pkg/helm/internal/util: refactored from pkg/helm
