+++ helm-app-operator/pkg/helm/controller/reconcile.go
@@ -27,6 +27,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/operator-framework/helm-app-operator-kit/helm-app-operator/pkg/helm"
+	"github.com/operator-framework/helm-app-operator-kit/helm-app-operator/pkg/helm/internal/util"
 )
 
 type helmOperatorReconciler struct {
@@ -45,21 +46,21 @@ func (r *helmOperatorReconciler) Reconcile(request reconcile.Request) (reconcile
 	o.SetGroupVersionKind(r.GVK)
 	o.SetNamespace(request.Namespace)
 	o.SetName(request.Name)
-	logrus.Debugf("Processing %s", helm.ResourceString(o))
+	logrus.Debugf("Processing %s", util.ResourceString(o))
 
 	err := r.Client.Get(context.TODO(), request.NamespacedName, o)
 	if apierrors.IsNotFound(err) {
 		return reconcile.Result{}, nil
 	}
 	if err != nil {
-		logrus.Errorf("failed to lookup %s: %s", helm.ResourceString(o), err)
+		logrus.Errorf("failed to lookup %s: %s", util.ResourceString(o), err)
 		return reconcile.Result{}, err
 	}
 
 	deleted := o.GetDeletionTimestamp() != nil
 	pendingFinalizers := o.GetFinalizers()
 	if !deleted && !contains(pendingFinalizers, finalizer) {
-		logrus.Debugf("Adding finalizer \"%s\" to %s", finalizer, helm.ResourceString(o))
+		logrus.Debugf("Adding finalizer \"%s\" to %s", finalizer, util.ResourceString(o))
 		finalizers := append(pendingFinalizers, finalizer)
 		o.SetFinalizers(finalizers)
 		err := r.Client.Update(context.TODO(), o)
@@ -67,7 +68,7 @@ func (r *helmOperatorReconciler) Reconcile(request reconcile.Request) (reconcile
 	}
 	if deleted {
 		if !contains(pendingFinalizers, finalizer) {
-			logrus.Infof("Resource %s is terminated, skipping reconciliation", helm.ResourceString(o))
+			logrus.Infof("Resource %s is terminated, skipping reconciliation", util.ResourceString(o))
 			return reconcile.Result{}, nil
 		}
 
@@ -89,14 +90,14 @@ func (r *helmOperatorReconciler) Reconcile(request reconcile.Request) (reconcile
 
 	updatedResource, needsUpdate, err := r.Installer.ReconcileRelease(o)
 	if err != nil {
-		logrus.Errorf("failed to reconcile release for %s: %s", helm.ResourceString(o), err)
+		logrus.Errorf("failed to reconcile release for %s: %s", util.ResourceString(o), err)
 		return reconcile.Result{}, err
 	}
 
 	if needsUpdate {
 		err = r.Client.Update(context.TODO(), updatedResource)
 		if err != nil {
-			logrus.Errorf("failed to update resource status for %s: %s", helm.ResourceString(o), err)
+			logrus.Errorf("failed to update resource status for %s: %s", util.ResourceString(o), err)
 			return reconcile.Result{}, err
 		}
 	}
